+++ scripts/main.tsx
@@ -13,18 +13,13 @@ import {NotFound} from "./components/NotFound";
 import {StorePicker} from "./components/StorePicker";
 import {Header} from "./components/Header";
 import {Fish} from "./components/Fish";
+import {Order} from "./components/Order";
 
 // Import Interfaces
-import {FishDataProps, FishObject} from "./interfaces";
+import {FishDataProps, FishObject, OrderProps} from "./interfaces";
 
 // Stylus
 import "../css/style.styl";
-
-interface OrderProps {
-  fishes: Object;
-  order: Object;
-  removeFromOrder(key: string);
-}
 
 interface InventoryProps {
   addFish(fish: FishObject);
@@ -40,14 +35,6 @@ interface UpdateFishProps {
   fish: FishObject;
   updateFish(key: string, attr: string, value: string | number);
   removeFish(key: string);
-}
-
-interface FishOrderProps {
-  key: string;
-  index: string;
-  fish: FishObject;
-  count: number;
-  removeFromOrder(key: string);
 }
 
 interface AddFishProps {
@@ -206,94 +193,6 @@ class AddFishForm extends React.Component<AddFishProps, any> {
         <input type="text" ref="image" placeholder="URL to Image" />
         <button type="submit">+ Add Item </button>
       </form>
-    );
-  }
-}
-
-/**
- * Order Container
- */
-class Order extends React.Component<OrderProps, any> {
-  private renderOrders = ( orderIds: string[] ) => {
-    return orderIds.map((key) => {
-      let fishOrderProps: FishOrderProps = {
-        key: key,
-        index: key,
-        fish: this.props.fishes[key],
-        count: this.props.order[key],
-        removeFromOrder: this.props.removeFromOrder
-      };
-
-      return (
-        <FishOrder {...fishOrderProps}/>
-      );
-    });
-  };
-
-  render() {
-    let orderIds = Object.keys(this.props.order);
-    let total = orderIds.reduce((prevTotal, key) => {
-      let fish: FishObject = this.props.fishes[key];
-      let count: number = this.props.order[key];
-      let isAvailable = fish && fish.status === "available";
-
-      if(fish && isAvailable) {
-        return prevTotal + (count * fish.price || 0);
-      }
-      return prevTotal;
-    }, 0);
-
-    let fishFlyIn = {
-      component: "ul",
-      transitionName: "order",
-      transitionEnterTimeout: 500,
-      transitionLeaveTimeout: 500,
-      className: "order"
-    };
-
-    return (
-      <div className="order-wrap">
-        <h2 className="order-title">Your Order</h2>
-        <CSSTransitionGroup {...fishFlyIn}>
-          {this.renderOrders(orderIds)}
-          <li className="total">
-            <strong>Total:</strong>
-            {h.formatPrice(total)}
-          </li>
-        </CSSTransitionGroup>
-      </div>
-    );
-  }
-}
-
-class FishOrder extends React.Component<FishOrderProps, any> {
-  private removeFromOrder = () => {
-    this.props.removeFromOrder(this.props.index);
-  };
-
-  render() {
-    let fish = this.props.fish;
-    let count = this.props.count;
-    let removeButton = <button onClick={this.removeFromOrder}>X</button>;
-    let countWheel = {
-      component: "span",
-      transitionName: "count",
-      className: "count",
-      transitionEnterTimeout: 250,
-      transitionLeaveTimeout: 250
-    };
-
-    if(!fish) {
-      return <li key={this.props.index}>Sorry, fish no longer available! {removeButton}</li>;
-    }
-    return (
-      <li key={this.props.index}>
-        <CSSTransitionGroup {...countWheel}>
-          <span key={count}>{count}</span>
-        </CSSTransitionGroup>
-        lbs {fish.name} {removeButton}
-        <span className="price">{h.formatPrice(count * fish.price)}</span>
-      </li>
     );
   }
 }
